Created some basics
