+++ crates/moss-git-hosting-provider/Cargo.toml
@@ -1,22 +1,23 @@
 [package]
-name = "moss_git"
+name = "moss_git_hosting_provider"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+moss_git.workspace = true
+
 tokio.workspace = true
 anyhow.workspace = true
 parking_lot = { workspace = true, features = ["serde"] }
 git2.workspace = true
-# gix = {version =  "0.70.0"}
-
+url.workspace = true
 dotenv = "0.15.0"
 serde_json = "1.0.138"
-oauth2 = { workspace = true, features = ["reqwest-blocking"] } # Using blocking for easy testing here
+oauth2 = { workspace = true, features = [
+    "reqwest-blocking",
+] } # Using blocking for easy testing here
 reqwest = { workspace = true, features = ["blocking"] }
-serde = { version = "1.0.217", features = [
-    "derive",
-] }
+serde = { version = "1.0.217", features = ["derive"] }
 webbrowser = "1.0.3"
 iota_stronghold = "2.1.0"
 zeroize = "1.8.1"
