+++ crates/moss-git-hosting-provider/src/github.rs
@@ -1,0 +1,2 @@
+pub mod auth;
+pub mod client;
